pyproject.toml
[project]
name = "df-infra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "fastapi>=0.112.1",
    "json-stream>=2.3.3",
    "pillow>=11.3.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "tomlkit>=0.13.3",
    "uvicorn>=0.35.0",
]
